added ERM relationships to models

+++ backend/src/main/java/com/shifts/backend/repository/CalendarRepo.java
@@ -7,7 +7,7 @@ import com.shifts.backend.model.Calendar;
 
 //this interface is used to access the database with the help of JPA
 @Repository
-public interface CalendarRepo extends JpaRepository<Calendar, Integer> {
+public interface CalendarRepo extends JpaRepository<Calendar, Long> {
 
     
 }
