+++ backend/src/main/java/com/shifts/backend/service/service/AvailabilityService.java
@@ -8,8 +8,8 @@ import com.shifts.backend.model.Availability;
 public interface AvailabilityService {
     Availability saveAvailability(Availability availability);
     List<Availability> getAllAvailabilities();
-    Availability getAvailabilityById(int id);
-    Availability getAvailabilityByEmployeeId(int employeeId);
+    Availability getAvailabilityById(Long id);
+    Availability getAvailabilityByEmployeeId(Long employeeId);
     Availability updateAvailability(Availability availability);
-    void deleteAvailability(int id);
+    void deleteAvailability(Long id);
 }
