+++ backend/src/main/java/com/shifts/backend/service/service/TimeBlockService.java
@@ -7,9 +7,9 @@ import com.shifts.backend.model.TimeBlock;
 public interface TimeBlockService {
     TimeBlockService saveTimeBlock(TimeBlock timeBlock);
     List<TimeBlock> getAllTimeBlocks();
-    TimeBlock getTimeBlockById(int id);
-    List<TimeBlock> getAllTimeBlocksByCalendarId(int calendarId);
+    TimeBlock getTimeBlockById(Long id);
+    List<TimeBlock> getAllTimeBlocksByCalendarId(Long calendarId);
     TimeBlock updateTimeBlock(TimeBlock timeBlock);
-    void deleteTimeBlock(int id);
+    void deleteTimeBlock(Long id);
 
 }
